[Cloud storage] Handling jsTree operations (server-side)

+++ application/controllers/Storagetree.php
@@ -5,6 +5,8 @@ class Storagetree extends CI_Controller {
 
 	public function index()
 	{
+		$this->load->helper('tree');
+
 		$operation = $this->input->get('operation');
 
 		try
@@ -17,8 +19,13 @@ class Storagetree extends CI_Controller {
 					$parent_uri = $this->input->post('parent_uri');
 					$text = $this->input->post('text');
 
+					if (empty($parent_uri) OR empty($text))
+					{
+						throw new Exception('Cannot create directory');
+					}
+
 					$parent_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $parent_uri);
-					$dir_path = $parent_path.rawurlencode($text);
+					$dir_path = path_encode($parent_path.$text, '/');
 
 					exec("mkdir {$dir_path}", $output, $status);
 					if ($status !== 0)
@@ -34,9 +41,14 @@ class Storagetree extends CI_Controller {
 					$new_text = $this->input->post('new_text');
 					$old_text = $this->input->post('old_text');
 
+					if (empty($parent_uri) OR empty($new_text) OR empty($old_text))
+					{
+						throw new Exception('Cannot rename directory');
+					}
+
 					$parent_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $parent_uri);
-					$new_path = $parent_path.rawurlencode($new_text);
-					$old_path = $parent_path.rawurlencode($old_text);
+					$new_path = path_encode($parent_path.$new_text, '/');
+					$old_path = path_encode($parent_path.$old_text, '/');
 
 					exec("mv {$old_path} {$new_path}", $output, $status);
 					if ($status !== 0)
@@ -50,7 +62,13 @@ class Storagetree extends CI_Controller {
 				case 'delete_node':
 					$uri = $this->input->post('uri');
 
+					if (empty($uri))
+					{
+						throw new Exception('Cannot delete directory');
+					}
+
 					$dir_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $uri);
+					$dir_path = path_encode($dir_path, '/');
 
 					exec("rm -r {$dir_path}", $output, $status);
 					if ($status !== 0)
@@ -66,8 +84,15 @@ class Storagetree extends CI_Controller {
 					$old_parent_uri = $this->input->post('old_parent_uri');
 					$text = $this->input->post('text');
 
-					$new_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $new_parent_uri).rawurlencode($text);
-					$old_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $old_parent_uri).rawurlencode($text);
+					if (empty($new_parent_uri) OR empty($old_parent_uri) OR empty($text))
+					{
+						throw new Exception('Cannot move directory');
+					}
+
+					$new_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $new_parent_uri).$text;
+					$new_path = path_encode($new_path, '/');
+					$old_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $old_parent_uri).$text;
+					$old_path = path_encode($old_path, '/');
 
 					exec("mv {$old_path} {$new_path}", $output, $status);
 					if ($status !== 0)
@@ -83,8 +108,15 @@ class Storagetree extends CI_Controller {
 					$old_parent_uri = $this->input->post('old_parent_uri');
 					$text = $this->input->post('text');
 
-					$new_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $new_parent_uri).rawurlencode($text);
-					$old_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $old_parent_uri).rawurlencode($text);
+					if (empty($new_parent_uri) OR empty($old_parent_uri) OR empty($text))
+					{
+						throw new Exception('Cannot copy directory');
+					}
+
+					$new_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $new_parent_uri).$text;
+					$new_path = path_encode($new_path, '/');
+					$old_path = STORAGEPATH.preg_replace('/^\/'.STORAGEURI.'/', '', $old_parent_uri).$text;
+					$old_path = path_encode($old_path, '/');
 
 					exec("cp -r {$old_path} {$new_path}", $output, $status);
 					if ($status !== 0)
